requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-multipart>=0.0.18,<1
python-dotenv>=1.1,<2
openai>=1.60,<2
pyyaml>=6.0,<7
